Refactor to fully support Factory creation & destruction
We have to jump through some hoops in order to fully make use of a
Factory operating on the generic Node type, including enforcing the
implementation of a `deinit` "method" on each *concrete* type to
destroy the heap allocation of that type instance.

It's a bit of boilerplate for each new node type, but I can't think of a
better solution that allows for an API dealing with only pointers to the
"base" Node type.