Refactor to support a global Context
It is assumed that users will create a global Context instance before
loading any Trees or otherwise creating any Nodes.

The Tree, while storing the global Context, does not take ownership
of it - the user retains ownership of the Context.

The Factory will use the single provided Context for all nodes it
creates while creating a new Tree.